Working on JesusRestaurant.

+++ src/simcity/jesusrestaurant/interfaces/JesusCashier.java
@@ -4,8 +4,8 @@ public interface JesusCashier {
 	
 	public abstract void msgComputeCheck (JesusWaiter w, JesusCustomer c, String foodItem, String name);
 
-	public abstract void msgCustomerPayment (JesusCustomer c, double money, String name);
+	public abstract void msgCustomerPayment (JesusCustomer c, int money, String name);
 	
-	public abstract void msgHereIsBill(JesusMarket m, double amount, int id);
+	public abstract void msgHereIsBill(JesusMarket m, int amount, int id);
 
 }
